Move follower to the correct package
